SW-4761 Create NurseriesRouter and move all nurseries related views (that exist within /nurseries path) and sub components to scenes directory
